consolidate grant and application schemas to avoid duplicate metadata
